# api/swagger.yaml
openapi: 3.0.3
info:
  title: SkopeoUI API
  version: 1.0.0
  description: SkopeoUI 后端 API 文档
servers:
  - url: localhost:8080
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
paths:
  /v1/once:
    post:
      tags:
        - once
      summary: 创建一次性任务
      operationId: CreateOnce
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Once'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  source:
                    type: string
                  dest:
                    type: string
        '400':
          description: 请求体解析失败
    delete:
      tags:
        - once
      summary: 删除一次性任务
      operationId: DeleteOnce
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: string
                example: "ok"
        '400':
          description: id 为空
  /v1/once/log:
    post:
      tags:
        - once
      summary: 获取一次性任务的日志流（SSE）
      operationId: GetOnceLog
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE 流（text/event-stream）
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          description: id 为空
        '404':
          description: 任务未找到
...
components:
  schemas:
...
    Once:
      type: object
      properties:
        source:
          type: string
          example: "docker.io/library/nginx:latest"
        destination:
          type: string
          example: "quay.io/myorg/nginx:latest"